Re-organize structrue
