+++ src/main/java/sypztep/dominatus/common/system/level/benefit/DefaultBenefitCalculator.java
@@ -1,4 +1,4 @@
-package sypztep.dominatus.common.util.level.benefit;
+package sypztep.dominatus.common.system.level.benefit;
 
 class DefaultBenefitCalculator implements BenefitCalculator {
     @Override
